requirements.txt
pandas
opencv-python-headless
Flask
camelot-py[cv]
gunicorn